\name{tennisratingscpp-package}
\alias{tennisratingscpp-package}
\alias{tennisratingscpp}
\docType{package}
\title{
  Fast C++ based calculation of Tennis Ratings
}
\description{
  This package contains functions written in C++ to quickly calculate ratings,
  such as player's serve and return elo ratings updated at each service point.
}
\author{
  Martin Ingram <martin.ingram@gmail.com>
  Maintainer: Martin Ingram <martin.ingram@gmail.com>
}
\keyword{ package }
